Cleaned old files, HPSMA backtest now runs.
